+++ CHANGELOG.md
@@ -4,7 +4,7 @@ __BACKWARDS INCOMPATIBILITIES:__
   * Client must be run as user with ability to call mount syscall
 
 IMPROVEMENTS:
-  * core: Introduce Constructor jobs and Dispatch command/API [GH-2128]
+  * core: Introduce Parameterized Jobs and Dispatch command/API [GH-2128]
   * core: Cancel blocked evals upon successful one for job [GH-2155]
   * cli: Defaulting to showing allocations which belong to currently registered
     job [GH-2032]